Widen the TTL column, match template replies to the quoted notice, quote timeline entries on mobile
Codex review on #903 found three gaps; each verified against the code
and fixed here.

- deal_update_recipients.link_ttl_ms is BIGINT: the 30-day choice is
  2,592,000,000 ms, past INTEGER's ceiling, so every recipient insert
  under that choice failed after the link or the send had already
  gone out. The additive migration now creates it as BIGINT and the
  held migration alters the already-created column.
- A buyer's reply to a purchase-progress notice is matched to the
  message it quotes (Meta's context.id → our messages row → the
  recipient row that carries it), never to the newest notice for the
  contact; with no quoted message the handler stands aside and the
  closing-nudge handler runs.
- The mobile composer offers the visibility-filtered timeline entries
  to quote, as the web one does, and submits them.

+++ supabase/migrations/20260918050000_transaction_workspace_updates.sql
@@ -115,7 +115,7 @@ CREATE TABLE IF NOT EXISTS deal_update_recipients (
     CHECK (status IN ('pending', 'sent', 'failed')),
   message_id UUID REFERENCES messages(id) ON DELETE SET NULL,
   failed_reason TEXT,
-  link_ttl_ms INTEGER,
+  link_ttl_ms BIGINT,
   link_otp_required BOOLEAN NOT NULL DEFAULT FALSE,
   sent_at TIMESTAMPTZ,
   link_delivered_at TIMESTAMPTZ,
